[refactor] Correct package path
