+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/component/MarkdownText.kt
@@ -40,7 +40,7 @@ import world.hachimi.app.ui.design.HachimiTheme
 import world.hachimi.app.ui.design.components.LocalContentColor
 import world.hachimi.app.ui.design.components.LocalTextStyle
 import world.hachimi.app.ui.design.components.Text
-import world.hachimi.app.util.fillMaxWidthIn
+import world.hachimi.app.ui.util.fillMaxWidthIn
 
 /**
  * Warning: Do not touch these fucking codes, it's generated by GPT-5.2
