+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/contributor/ContributorCenterRouteScreen.kt
@@ -26,8 +26,8 @@ import world.hachimi.app.ui.component.ReloadPage
 import world.hachimi.app.ui.creation.publish.PublishScreen
 import world.hachimi.app.ui.design.components.Button
 import world.hachimi.app.ui.design.components.Text
-import world.hachimi.app.util.AdaptiveScreenMargin
-import world.hachimi.app.util.fillMaxWidthIn
+import world.hachimi.app.ui.util.AdaptiveScreenMargin
+import world.hachimi.app.ui.util.fillMaxWidthIn
 
 @Composable
 fun ContributorCenterScreen(child: Route.Root.ContributorCenter) {
