+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/home/components/AdaptivePullToRefreshBox.kt
@@ -7,7 +7,7 @@ import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.Dp
-import world.hachimi.app.util.WindowSize
+import world.hachimi.app.ui.util.WindowSize
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
