+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/playlist/components/Header.kt
@@ -49,7 +49,7 @@ import world.hachimi.app.ui.design.components.Icon
 import world.hachimi.app.ui.design.components.LocalContentColor
 import world.hachimi.app.ui.design.components.Text
 import world.hachimi.app.ui.theme.PreviewTheme
-import world.hachimi.app.util.fillMaxWidthIn
+import world.hachimi.app.ui.util.fillMaxWidthIn
 import kotlin.time.Clock
 import kotlin.time.Instant
