+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/recentplay/RecentPlayScreen.kt
@@ -46,9 +46,9 @@ import world.hachimi.app.ui.component.ReloadPage
 import world.hachimi.app.ui.design.components.Surface
 import world.hachimi.app.ui.design.components.Text
 import world.hachimi.app.ui.theme.PreviewTheme
-import world.hachimi.app.util.AdaptiveScreenMargin
+import world.hachimi.app.ui.util.AdaptiveScreenMargin
+import world.hachimi.app.ui.util.contentPaddingForMaxWidth
 import world.hachimi.app.util.YMDHM
-import world.hachimi.app.util.contentPaddingForMaxWidth
 import world.hachimi.app.util.formatTime
 import kotlin.time.Instant
