+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/settings/SettingsScreen.kt
@@ -74,8 +74,8 @@ import world.hachimi.app.ui.design.components.Select
 import world.hachimi.app.ui.design.components.Switcher
 import world.hachimi.app.ui.design.components.Text
 import world.hachimi.app.ui.design.components.TextButton
-import world.hachimi.app.util.AdaptiveScreenMargin
-import world.hachimi.app.util.fillMaxWidthIn
+import world.hachimi.app.ui.util.AdaptiveScreenMargin
+import world.hachimi.app.ui.util.fillMaxWidthIn
 
 @Composable
 fun SettingsScreen(
