+++ composeApp/src/commonMain/kotlin/world/hachimi/app/ui/userspace/EditProfileScreen.kt
@@ -87,9 +87,9 @@ import world.hachimi.app.ui.design.components.Text
 import world.hachimi.app.ui.design.components.TextButton
 import world.hachimi.app.ui.design.components.TextField
 import world.hachimi.app.ui.design.components.ToggleButton
+import world.hachimi.app.ui.util.AdaptiveScreenMargin
 import world.hachimi.app.ui.util.PlatformIcons
-import world.hachimi.app.util.AdaptiveScreenMargin
-import world.hachimi.app.util.fillMaxWidthIn
+import world.hachimi.app.ui.util.fillMaxWidthIn
 
 @Composable
 fun EditProfileScreen(vm: EditProfileViewModel = koinViewModel()) {
